requirements.txt
Pyrogram
PyYAML
tgcrypto
pymongo
dnspython
